refactored lamdba implementation
